GN-4510: Extract defaults to separate file
