+++ app/components/simple-editor.js
@@ -2,14 +2,17 @@ import Component from '@glimmer/component';
 import { action } from '@ember/object';
 import { tracked } from '@glimmer/tracking';
 import { inject as service } from '@ember/service';
-import merge from 'lodash.mergewith';
 
 import { Schema } from '@lblod/ember-rdfa-editor';
 
-import { linkPasteHandler } from '@lblod/ember-rdfa-editor/plugins/link';
 import {
-  tableOfContentsView,
+  link,
+  linkPasteHandler,
+  linkView,
+} from '@lblod/ember-rdfa-editor/plugins/link';
+import {
   table_of_contents,
+  tableOfContentsView,
 } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/table-of-contents-plugin/nodes';
 import { firefoxCursorFix } from '@lblod/ember-rdfa-editor/plugins/firefox-cursor-fix';
 import { lastKeyPressedPlugin } from '@lblod/ember-rdfa-editor/plugins/last-key-pressed';
@@ -23,8 +26,8 @@ import {
   underline,
 } from '@lblod/ember-rdfa-editor/plugins/text-style';
 import {
-  docWithConfig,
   block_rdfa,
+  docWithConfig,
   hard_break,
   horizontal_rule,
   invisible_rdfa,
@@ -73,7 +76,6 @@ import {
 import { citationPlugin } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/citation-plugin';
 
 import { roadsign_regulation } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/roadsign-regulation-plugin/nodes';
-import { link, linkView } from '@lblod/ember-rdfa-editor/plugins/link';
 import { highlight } from '@lblod/ember-rdfa-editor/plugins/highlight/marks/highlight';
 import { color } from '@lblod/ember-rdfa-editor/plugins/color/marks/color';
 
@@ -86,99 +88,22 @@ import {
   addressView,
   codelist,
   codelistView,
-  number,
-  numberView,
   location,
   locationView,
+  number,
+  numberView,
   text_variable,
   textVariableView,
 } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/variable-plugin/variables';
 
-const mergeCustomizer = (objValue, srcValue) => {
-  if (Array.isArray(objValue) && Array.isArray(srcValue)) {
-    return srcValue;
-  }
-};
-
-/**
- * If the user config is present, merge it with the default config.
- * Otherwise return the default config.
- * @param defaultConfig
- * @param userConfig
- * @returns {*}
- */
-const mergeConfigs = (defaultConfig, userConfig) => {
-  if (userConfig) {
-    return merge(defaultConfig, userConfig, mergeCustomizer);
-  }
-
-  return defaultConfig;
-};
-
-/**
- * @type {import('@lblod/ember-rdfa-editor-lblod-plugins/plugins/rdfa-date-plugin/index').DateOptions}
- */
-const defaultRdfaDatePluginConfig = (t) => ({
-  placeholder: {
-    insertDate: t('date-plugin.insert.date'),
-    insertDateTime: t('date-plugin.insert.datetime'),
-  },
-  formats: [
-    {
-      label: 'Short Date',
-      key: 'short',
-      dateFormat: 'dd/MM/yy',
-      dateTimeFormat: 'dd/MM/yy HH:mm',
-    },
-    {
-      label: 'Long Date',
-      key: 'long',
-      dateFormat: 'EEEE dd MMMM yyyy',
-      dateTimeFormat: 'PPPPp',
-    },
-  ],
-  allowCustomFormat: true,
-});
-
-/**
- * @type {import('@lblod/ember-rdfa-editor-lblod-plugins/plugins/citation-plugin').CitationPluginEmberComponentConfig}
- */
-const defaultCitationPluginConfig = {
-  type: 'ranges',
-  activeInRanges: (state) => [[0, state.doc.content.size]],
-  endpoint: '/codex/sparql',
-};
-
-/**
- * @type {import('@lblod/ember-rdfa-editor-lblod-plugins/plugins/roadsign-regulation-plugin').RoadsignRegulationPluginOptions}
- */
-const defaultRoadsignRegulationPluginConfig = {
-  endpoint: 'https://dev.roadsigns.lblod.info/sparql',
-  imageBaseUrl: 'https://register.mobiliteit.vlaanderen.be/',
-};
-
-/**
- * @type {import("@lblod/ember-rdfa-editor-lblod-plugins/components/variable-plugin/location/edit").LocationEditOptions}
- */
-const defaultLocationVariablePluginConfig = {
-  endpoint: 'https://dev.roadsigns.lblod.info/sparql',
-  zonalLocationCodelistUri:
-    'http://lblod.data.gift/concept-schemes/62331E6900730AE7B99DF7EF',
-  nonZonalLocationCodelistUri:
-    'http://lblod.data.gift/concept-schemes/62331FDD00730AE7B99DF7F2',
-};
-
-/**
- * @type {import("@lblod/ember-rdfa-editor-lblod-plugins/plugins/table-of-contents-plugin").TableOfContentsConfig}
- */
-const defaultTableOfContentsPluginConfig = [
-  {
-    nodeHierarchy: [
-      'title|chapter|section|subsection|article',
-      'structure_header|article_header',
-    ],
-  },
-];
+import {
+  defaultCitationPluginConfig,
+  defaultLocationVariablePluginConfig,
+  defaultRdfaDatePluginConfig,
+  defaultRoadsignRegulationPluginConfig,
+  defaultTableOfContentsPluginConfig,
+  mergeConfigs,
+} from '../config/defaults';
 
 export default class SimpleEditorComponent extends Component {
   @tracked controller;
